pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-character"
version = "0.1.8"
description = "A Python package for creating interactive AI characters"
readme = "README.md"
authors = [{name = "Sid Uppal", email = "siddhu@gmail.com"}]
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "openai",
    "sounddevice",
    "numpy",
    "pygame",
    "pydub",
    "elevenlabs",
    "pyyaml",
    "opencv-python",
    "groq",
    "ollama",
    "scipy"
]
requires-python = ">=3.8"

[project.urls]
Homepage = "https://github.com/sidu/ai_character"
